requirements.txt
streamlit==1.27.1
Pillow==9.2.0
pandas==1.4.4
openpyxl==3.0.10